requirements.txt
rembg
pillow